refactor: naming from code review
matmul/_systolic.py and _fp64_accum.py behind the functions of the same name; scheme factories uppercase
like the Arithmetic ones (MXQUANT, MXGEMMINI, OCP_FP64, PASSTHROUGH); schedule.load/fixed count rows;
reducers take and return float32, checked in check_operands.

+++ mxq/matmul/_systolic.py
@@ -44,4 +44,4 @@ def systolic(P_A: torch.Tensor, X_A: torch.Tensor, P_B: torch.Tensor, X_B: torch
                 e, m = schedule[k % window]
                 S = arith.acc_add(S, p, e, m)
             C = arith.tile_add(C, S * scales)
-    return C
+    return C.to(torch.float32)
